Finished Data Tables

+++ src/server/routers/project.ts
@@ -1,8 +1,14 @@
+import { Stage } from "@prisma/client";
 import { z } from "zod";
 
+import { stageCheck } from "@/lib/utils/permissions/stage-check";
 import { instanceParamsSchema } from "@/lib/validations/params";
 
-import { createTRPCRouter, protectedProcedure } from "@/server/trpc";
+import {
+  createTRPCRouter,
+  protectedProcedure,
+  stageAwareProcedure,
+} from "@/server/trpc";
 
 export const projectRouter = createTRPCRouter({
   getTableData: protectedProcedure
@@ -22,6 +28,12 @@ export const projectRouter = createTRPCRouter({
             id: true,
             title: true,
             description: true,
+            flagOnProjects: {
+              select: { flag: { select: { id: true, title: true } } },
+            },
+            tagOnProject: {
+              select: { tag: { select: { id: true, title: true } } },
+            },
             supervisor: {
               select: { user: { select: { name: true, id: true } } },
             },
@@ -49,7 +61,7 @@ export const projectRouter = createTRPCRouter({
       });
     }),
 
-  delete: protectedProcedure
+  delete: stageAwareProcedure
     .input(z.object({ params: instanceParamsSchema, projectId: z.string() }))
     .mutation(
       async ({
@@ -59,6 +71,8 @@ export const projectRouter = createTRPCRouter({
           projectId,
         },
       }) => {
+        if (stageCheck(ctx.stage, Stage.PROJECT_ALLOCATION)) return;
+
         await ctx.db.project.delete({
           where: {
             allocationGroupId: group,
@@ -70,7 +84,7 @@ export const projectRouter = createTRPCRouter({
       },
     ),
 
-  deleteAll: protectedProcedure
+  deleteAll: stageAwareProcedure
     .input(z.object({ params: instanceParamsSchema }))
     .mutation(
       async ({
@@ -79,6 +93,8 @@ export const projectRouter = createTRPCRouter({
           params: { group, subGroup, instance },
         },
       }) => {
+        if (stageCheck(ctx.stage, Stage.PROJECT_ALLOCATION)) return;
+
         await ctx.db.project.deleteMany({
           where: {
             allocationGroupId: group,
@@ -86,6 +102,52 @@ export const projectRouter = createTRPCRouter({
             allocationInstanceId: instance,
           },
         });
+      },
+    ),
+
+  details: protectedProcedure
+    .input(z.object({ params: instanceParamsSchema }))
+    .query(
+      async ({
+        ctx,
+        input: {
+          params: { group, subGroup, instance },
+        },
+      }) => {
+        const allFlags = await ctx.db.flag.findMany({
+          where: {
+            allocationGroupId: group,
+            allocationSubGroupId: subGroup,
+            allocationInstanceId: instance,
+          },
+          select: {
+            id: true,
+            title: true,
+            flagOnProjects: true,
+          },
+        });
+
+        const allTags = await ctx.db.tag.findMany({
+          where: {
+            allocationGroupId: group,
+            allocationSubGroupId: subGroup,
+            allocationInstanceId: instance,
+          },
+          select: {
+            id: true,
+            title: true,
+            tagOnProject: true,
+          },
+        });
+
+        return {
+          flags: allFlags
+            .filter((f) => f.flagOnProjects.length !== 0)
+            .map((e) => ({ id: e.id, title: e.title })),
+          tags: allTags
+            .filter((t) => t.tagOnProject.length !== 0)
+            .map((e) => ({ id: e.id, title: e.title })),
+        };
       },
     ),
 });
